Revert the upgrade from tailwind 3 to 4
This reverts most of commit 64ce2d6a0755369a3c64820a48c565d7ea6ed6fd.

Filament 3 does not support tailwind 4 and it was
being too difficult to work with on the backend.
It would eventually be fixable but I don't have
the time to fix it and was honestly not enjoying
working on this anymore after that.